backport migration EE code structure
